Map presets to pedalboard presets (#69)
Currently the zynthian presets are mapped to mod-ui effect presets. Mod-ui also has a concept of a pedalboard preset which maps better to zynthians model of a preset.

This change begins the preset list with a list of pedalboard presets, then displays effect presets after the divider. The effect presets are now prepended with the effect name so that the effect preset list isn't just a list of presets called 'Default'.