On global playback start, disable any Patterns not assigned to a Track
This ensures that people don't end up suddenly hearing notes they don't
know where are coming from when on e.g. Looper. They will still be able
to force them back on if they so desire by going to drumatique and
hitting play in the sidebar there, or just enabling the pattern
explicitly in the popup, but this at least ensures less surprises.