Add visualPressAndHold on PlayGridButton, and then show it too
This adds a little rectangle which fades in and fills up beside the
PlayGridButton when held down, if the new property is set to true. This
intends to make it obvious and discoverable that something might
actually happen if you hold the button down rather than just tapping it,
without it also being so in your face that it's distracting.