Add ability in Drumatique to listen for multiple notes, and to set them on sequences
This adds in the ability to listen to multiple notes (basically,
chords), and allows you to also set/unset them on patterns in a sequence
the same way you would single notes. The logic for set/unset is
essentially the same (that is, is there a note from the bunch in that
position then that note is removed, and if not then all the notes in the
bunch is added, with their associated velocities). Press and hold the
Note: button to start listening, press it again to stop and select those
notes, or press and hold to discard the current list and restart
listening. All other interaction remains the same.