Attempt to remove that refreshSteps hack, and add Get New Sequences action
The former might have a fair bit of influence on just how fast
drumatique's UI is, but also i am unsure if it is in fact still working
right. From what i can tell it is, but at the same time, we will want to
be sure. Also, potentially, there is more that can be done if this
/does/ fix everything (basically, thinking we maybe don't need to
recreate the components, and instead we can just change their
contents... maybe)