Switch to using a static number for pattern popup and bar display
This is less correct than the previous code, but as delegate
recycling/caching isn't really a thing in Qt 5.12, we will need to work
around that. A lot of the code sort of assumed something similar anyway
already, so it shouldn't be a huge upheaval, but it will need some
thorough testing to make sure everything's in sync.