drumatique: step setting delegate now handles interpreted values gracefully
The longer version of the above is that we have an actual default value
(which can be either an integer or undefined, depending on the parameter),
and that actual default value should be /displayed/ as something else in
the UI, and further operate like a real value for interaction purposes.
That means we have to add handling of that into the locations where that
value is set, and also where it is displayed.

So, we now force set the real value to the default whenever we reach that
interpreted value when navigating through the values with either the
slide control, or either of the - or + buttons. It was already
functionally taken into account for display, but a minor update to the
logic to centralise some checks is also included.

Finally, minor layout adjustment