Rejig how enabled/disabled is synced between track/pattern
This fixes the weird behaviour we had previously, caused by a mismatch
in how the scene/track/clip/sequence/pattern interrelations work. This
was due to a situation from the long-ago, before scenes had properly
entered the world, and we instead had a concept of multiple clips per
track. Since those clips are now equivalent to scenes, rather than being
equivalent to a bank in a pattern, the old logic here was badly broken
and caused inconsistencies. So, now the logic is polished up
considerably, and should work much better.