Sync logic: Rework to track/part/scene inverse hierarchy, force forward-sync
This is a major reworking of how the synchronisation logic functions,
from synchronising bi-directionally, to a system which only forwards
settings from the track/part/scene object hierarchy into the
sequence/pattern hierarchy.

Furthermore, it represents a personal eureka moment, as can also be seen
in the rewording of the comment describing the ideas behind the two
structures. Documentation is going to be so much nicer to write given
this understanding, makes me glad ;) Took a bit of a while to figure it
out, though, but oh well, better late than never, and definitely also
better now than later, and certainly before any non-alpha releases!