Rework enabled logic to sit in track/clip (allowing for better magic)
The logic here is: If we're enabling a clip, check in the clip's track
whether we allow multi part playback. If we don't, disable all the other
parts (clips) on that scene/track, and if we do, leave them alone. This
is also supposed to be changed when switching away from a multi part
allowing mode, but that seems to be not working right now. Better than
it was, though.