zl.track, zl.clip: exception-guard the deserialisation function
This seems to be the prime cause of our various "wtf" moments with
deserialising songs. Hopefully things should work much more consistently
with this in place (not that those errors should occur to start with,
but...)

This also moves the selectedPart restoration to the end, so we're not
trying to set current a part that hasn't been loaded yet.