zynthian: Disable hardware handling (in/out) in autorouter
This now depends on ZLRouter piping events to the autorouter, which will
happen automagically if you are using Zynthbox. It will basically make
Zynthian running just as Zynthian unusable, and we are likely at some
point going to need an option in Zynthian to do this, rather than this
hard-coded disabling.