zl.clips_model: Update samples prop based on track data
This implements the initial logic we came up with, where in
trig/auto-full mode it will assign a sample to each pattern explicitly.
The automagic will get in the way if we implement custom picking, so we
probably want this to be slightly more clever, but we can get to that
when it gets to that point.