drumatique: Fix part-related todo, fix up sound-descriptioning, fix parts popup logic
The parts related todo meant we'd not correctly set the part when
picking a pattern from the menu, leaving us badly out of sync.

The sounds label for patterns now says something much more useful for
various things (for sample-trig it says what samples are available for
the pattern, for external it says what channel it'll use)

Parts popup is more aligned with how everything else wants to work.
Still some wording to wiggle, but it works much more predictably now.