PlayGrid sync: Fix a bunch of track->sequence hierarchy syncing issues
We're now using Bindings for a bunch of the things where that
makes sense (enabled, noteDestination, and externalMidiChannel), and
we're using the Bindings item's when ability to make sure we're updated
when we adopt a track. It's slightly overzealous, but also it's what we
need.

We're now also setting the part when doing the sequence setByTrack
thing, which we didn't before. Much confusion, very bad. Much better
now.

There's likely more stuff wants doing in there, and some stuff that's
just outright redundant, but it feels like good progress.