StatusInfo: Use new ZLRouter info (via PlayGridManager) to show hardware notes
This also slightly rejigs how the notes are displayed, otherwise it just
looks overly cramped or just outright not useful. Shows the first note
in the list, and how many more if there are more. This happens for both
hardware in notes, hardware out notes, and the existing "all notes"
thing.