SynthSetupPage: Suppress a binding loop
This is a bit roudnabout, however: Calling the selector_list for preset
and bank causes that list to be regenerated, which in turn causes .count
to change, and then all the bindings end up updating again. So, to
short-stop that, we need to add this pile of stuff. It also results in
the whole thing being faster and lighter, so that's always nice.

Optionally, we might want to add a .count to the component itself rather
than its selector_list, which then (one would hope) would not result in
this recursive update behaviour.