drumatique StepSettings: Switch duration order, add pitch change func
The former is the first step and apparently it's meant to be decoupled
from the pattern note length, so that's something that will need
discussing more deeply.

The latter is perhaps not the best visual location for it, but not sure
where else to put it. The functionality requires updated components.