drumatique: Block direct-track change, rejig up/down/left/right nav
We don't want to select tracks with the track buttons directly (we still
use them for "clear the pattern" stuff with the back key pressed for
now), so consume all the track buttony things, not just for back
buttons.

Switch the arrowkey navigation:

up/down changes the active bar
left/right changes track

Also fixes being able to navigate to bar number -1 (not a thing).