ZL : Handle 5th under screen button to open context specific edit page
Change the fifth softbutton underneath screen from Settings to EDIT... that is when synth is and slot is active, edit that sound. if sampler is active, go to wav editor. if external track is active, dont do anything yet. if loop audio is active, go to wav editor also. if no sound or sample or loop has been chosen yet or you are on empty slot, show pick option like if pressing on an empty slot, synth or sample or file selector for loop