PlayGrid sync: Add notificationing for hw add/remove, wants new libzl
This causes a small notification to pop up when a hardware device is
plugged in or unplugged. Small thing, but because of the highly awkward
nature of the raspberry pi's usb stack (hardware or whatever), and it
failing to correctly handle a device showing up, it means we can now
more easily handle that, without having to just... plug/unplug and watch
the journal output, which isn't exactly user friendly. More user
friendly would be to not have a broken usb stack, but that's what we're
dealing with, so let's at least be slightly more graceful, i guess.

This doesn't break things if libzl isn't upgraded, and will only
complain about missing signals.