notesgrid: Fix notesgrid basics, and sort by-row scrolling
Switch to a new method for creating the rows, and split the logic for
row starts into its own thing. Much more stable, and considerably easier
to work with.

The notes grid has been broken for a fair while at this point, and took
some work to wrangle back into shape. It should, however, also yield a
better experience in general.