notesgrid: Don't transpose by default - confusing if key is C
This needs more work, as we basically want to be able to place a
specified base note in the bottom-left by-default (the song's key), and
then operate from that. So, we need to generate the stuff backwards as
well as forward, so that'll need doing. For now, though, this will have
to do.