drumatique: Add NoteSettings component, for editing notes on multiple steps
This is a modified version of the StepSettings component, but that one
operates on a single step, where this is supposed to operate on a
per-bar basis, with filtering, and so is a bit more involved. We might
eventually want to also filter on steps the way we do on notes, but for
now, this will do.