drumatique: Allow negative "delay" settings, so we can position notes before their step
This, paired with new code in PatternModel, should allow us to position
notes in the past by one step position. Paired with the work done last
week to make playback single-step progression instead of explicit scheduling
ahead, this should allow us to have more control over what notes are
fired when. We would technically be able to also fire notes further
back, but the lookahead does become more involved as we increase the
amount. However, we already know that at least eight is barely
noticeable, so we could probably get away with quite a lot.
Experimentation will show whether we want to.