gui effects types, layer effects, selector handle empties more gracefully
While working on the autoconnection stuff for SamplerSynth, ran into
these not really being too happy if there was nothing in their lists and
such, so... some simple checks for now. There's more to be done to make
it work more gooder, but it now doesn't get quite so angry.