autoconnect: If there's no track fx, connect track to global fx, otherwise playback
In slightly longer form:
When connecting the samplersynth outputs, first check whether the track
has any effects associated with it. If it does, connect to those. If it
doesn't, check if there are any global effects. If there are, connect to
them. If there's no global effects set, connect directly to system
playback.