zynautoconnect : Always check if layer in chainedSound exists.
Sometimes when choosing synth a layer might be assigned to chainedSounds but layer might not have been created as user might have backed out. In those cases chainedSound might contain a layer that is not -1 but doesnt not exist either.

This specific case was causing load_snapshot to stop working and hence loading a sketch was getting stuck.