zynautoconnect: Work around an occasional hiccup
It happens rarely, far as we can tell, but occasionally the engines
won't be completely ready yet, and their jack port name won't be filled
out by the time we need it. So, handle that everywhere we need it, and
postpone the call a short while, at which point it should hopefully be
ready.