Multi track recorder: Implement the basic functionality
The first multi-track recording has been made. We will need to do the
song duration and whatnot a little differently to this, and the leadin
and fadeout durations need to be a little more automagic as well, but
the default-toggle for tracks and such seems to be sensible enough, and
the recording does, in fact, work. So, onwards!