ExternalRecordingDestinationClip : Do not use tracksModel directly in a repeater and use a static count instead
This was causing setting selectedPart of track[0] to get delayed by around 400ms when clicking on a part clip cell in partBar. The delay problem was only happening for track[0] and none of the other tracks.

Still any models of any repeater like components should be static numbers to avoid any delays when children are re-rendered