autoconnect: Use the autoconnect thread to postpone the retry
There's little reason to use a timer here, when we already have a thread
that will do what we want, if it's just told to do it. So, set a boolean
that gets handled by the thread call instead, and as a bonus we end up
not clashing (and crashing).