CUIA: Handle cuia in main window, forward to vkbd and other dlgs
This adds a small extra layer so we can explicitly handle cuia in the
main window's code as well, giving us a central place in qml to handle
such things, essentially mirroring the central nature of the cuia system
in python. We could arguably remove some of the existing logic in favour
of this, but we can get to refactoring that later if we want to. For
now, this ensures we're able to close the virtual keyboard and whatnot
as we want to with the hardware buttons.