zynthian_qt_gui : A Hacky way to set initial values of knobs correctly.
Keep calling set_selector and set options forcefully until booting is completed. This is done because if set_selector is not called multiple times initial value is always wrong no matter what is done. Until a proper solution is found, this have to do.

See zynthian_gui_controller `config` method for a check before logging when booting is in progress. If booting is not yet complete log is disabled as it becomes very noisy. When a fix is found, remove the check before logging from zynthian_gui_controller::config