drumatique: Add live recording to Play/Pause button as press-and-hold
This probably is not really the perfect location (or, rather, that
play/pause button probably actually wants to be just outright turned
into a record toggle for live-recording patterns, as we already have
start/stop controls right there in the main toolbar, and this feels kind
of like a leftover). For now, though, it works as a way to make the live
recording stuff happen.