osd: Much more pleasant interaction
This introduces a similar slidey thing as used in the step parameter
delegate in drumatique. It also changes the logic a little bit so that
if we're holding a finger on the osd, and the value changes, the osd
will still stay up. Minor layout adjustments as well, and all in all it
feels much nicer now.