osd: Add a visualZero concept to the python osd object
The logic here is to allow the user of the api to offset the zero point
by some amount. By default, this will be set to the start point, but as
an example, pan is from -1 to 1, and the visual zero point would be 0,
which will then allow us to visualise the left/right nature of it,
instead of showing it as just a ltr progress bar.