Fix jack connections logic, and only connect used ports
This is a fairly big chunk of work, and it may well have introduced
problems. It also, however, fixed problems that previously existed
(specifically, the audiolevels system playback client was connected to
far too many things), and it further drastically reduces the amount of
mixing jack is required to do, which should reduce the waiting around
it's having to then do.

This also is fairly important for the
diable-jack-processing-for-samplersynth-channel-on-non-sample-channel
logic, as without this, if there's a sample playing on the channel when
the channel is switched away from sample mode, we'll end up just
spitting out the same buffer over and over, which is kind of noisy.