Autoconnect: Adapt the global fx connections to libzl JackPassthrough client
Use the new GlobalFXPassthrough client in libzl to pass data through sound
from all the various sound producers to the global effect engines. The logic
here is that there are two engines, and having a pile of connections to both
of those is fairly expensive when compared to a single set of those to a client
that then (in the default configuration) just copies the data from one
buffer to another. That, at any rate, is the logic. Hotspot seems to
agree, and the graph looks more pleasant as well.