sketchpad.clip: Add slot and auto-stopping to play/stop functions
These are basically not used anywhere, but function just fine. So, as we
need something like it, and the intention of a "play" function would be
to play until the end anyway, without looping, let's make sure it's also
auto-stopped when it reaches the end. Should have no effect other than
fixing that minor issue, and making it available to qml (such as for
e.g. the recording location picker popup in the module panel)