ExternalRecordingDestinationClip: Actually implement the component
This adds the ability to pick a specific sample in a specific channel,
and then record directly into that. It understands how to record a
specific set of jack ports only, or the full system, and will also show
a preview of the selected sample, as well as allow you to play/stop that
sample.

This follows the current channel-sample selection, and will also set it
when you pick it in the grid.

Hopefully it should be decently optimised, but testing will be needed to
ensure it's working as intended.