main: Fix some record destination popup visual weirdness
Most important one probably the "didn't show up when first tapped and
had to tap three times to make the thing show up" thing, which... well,
not really entirely sure why that doesn't work, but QML Window is weird
at times, and i guess setting it visible just made it visible and didn't
actually put it on top. So, this does the trick very nicely, and reads
fine anyway.

Also had some trouble getting the tab bar picking thing to work, so this
is more a workaround than anything else, but it works pretty well, so
it's not a big loss anyway, and it seems to be faster as well, which is
kind of nice as well.

Finally, move the start/stop recording logic into the components
themselves, which is where that logic really wants to live anyway.