stepseq: Remove audio type quad, add samplevisualiser where appropriate
This firstly removes the channel audio type quad from both the pattern
popup, and from the settings bar (this really is something that should
all be done from the main sketchpad anyway, and here it's mostly noise,
and just kind of not really the most directly useful thing anyway).

It also introduces sample visualisation in a couple of places where it
makes sense:

For a sample-trig/slice channel, the sample is visualised in the sound
picker button in the pattern popup

For a sample-loop channel, the sample is visualised in place of the
pattern in the pattern popup, and in place of the drumpad steps in the
main view.

There's likely other situations where this would make sense, but these
already show the idea.