bt config: Add auto-retry (tries five times when connecting)
This works around the occasional situation where connecting to a bt
audio device takes longer or fails the first time and the alsa device
creation takes a little bit longer than usual. It's not quite perfect
(probably wants some kind of throttle added so it's not just a hammer),
but should work reasonably well at least.