main gui volume&fx setters/getters, now without stealing encoders
This adds setters and getters for the global volume, and the two global
effects (fx1 and fx2), with the intention of abstracting away the fact
they are currently delay and reverb respectively, as we eventually want
them to be configurable and the code will want to reflect that
abstraction.

The new setter/getters could be turned into properties, but for
now they just live there. Their addition also required changing the
existing setters to allow for not forcing stealing of the encoders that
are used to control them on the main screen. This is in preparation for
using their core functionality (the osd and the actual value settery) in
the playground page, coming momentarily.

This also puts in a queue to do the file storage thing for the current
task message. Doesn't actually fix the thing i hoped it'd fix, but it's
still better to avoid doing disk i/o in a property getter, so leaving it
in.