mixer: Store the volume as 0-100 locally, and push that to alsa
The result here is that we don't have to to two-way interpretations,
which is just a great way of making values jump around a bunch. It's a
bit simpler to read, and arguably it might be less "correct", as we're
storing a value that we don't /know/ might have changed in Alsa, but
also we are supposed to be the canonical value and alsa doesn't tell us
anyway. So, this works, and makes changing the value much smoother as
well. Calling it a win.