Add a stack to the dialog system, use in sequencer
This adds a way for us to have layers of dialogues. For example, you
might open a dialog and then show another popup to pick an item, and we
need to support that kind of interaction. As an example, in the
sequencer we've got the Pattern popup, which lets you pick what track
you're seeing in the sequencer, and that popup will in turn let you pick
a couple of things (such as the particular part of a channel, or show
the combinator), and we need to be able to easily close those. So, we
now have a stack, and it allows us to more easily handle that sort of
concept. There's other places where this would help, but for now this
shows that it works.