Song mode: Adapt a contextual song playback logic approach
alt+play plays in song mode, play plays in not song mode
Except for in the song manager, where the default is to play in song
mode and alt will play in not song mode.

Stopping playback will always revert to turning off song mode.

Simple logic is nice, and we no longer have a "mode" :)