cpu status check: Honour option, run in thread
With this check off, i am no longer seeing the spikes in processing
time, which leads me to think that the vcgencmd command does something
really nasty to the cpu... which is kind of weird, given it's supposed
to be used for checking its state, but perhaps there's interrupt stuff
in there or something? At any rate, with this check turned off, things
seem happier? More testing needed, but we would want to honour the
toggle anyway.

Also popped it into a thread, which was the first attempt at making it
lighter (as it used to be run every all the time a lot, which is overly
heavy anyway, and we don't need it that exact anyway, so it's run every
half a second when the toggle is enabled now).