cpu watchdog: Pull the overtemp/undervolt checker out of the main process
After further investigation, it turns out that it isn't the vcgencmd
command itself that's a problem, but rather how check_output works which
causes the entire process to stall for a few moments on a regular basis.
So, pull it out into a separate process, and communicate with that
through a fifo, hey presto stability /and/ return of the feature itself.
Nice. Thanks to Anupam for doing the tests that showed the above!