LED Config : Rework to update the led values on dependent value change instead of polling
This commit completely changes the way LED colors are updated. The existing implementation used a thread with an infinite loop to updated the LEDs every 0.05s which is pretty inefficient.

The current implementation makes use of the signals that are already available to update the led colors when the dependent values changes