stepseq stepsettings: Row-layoutify note info row, controls always visible
This equalises the design of the step settings popup with the changes in
the note settings popup. Would be kind of nice if we could just have one
for both purposes, but it's a bit awkward to trick the note settings one
to show a single step and also look sensible at the same time, so...
this will do.