zynthian_gui_layer : Do not aquire lock when removing layer and do not force autoconnect when removing layer
Explanation : Foo YC20 synth is causing massive amount of XRUNs and this is wreaking havoc in all other places. One of those places is here, i.e. removing a synth.

If there are massive number of XRUNs(CPU is busy), acquiring lock takes huge amount of time (somewhere between 20-50 seconds).
So when removing a synth if a lock is acquired, the removal call gets stuc for about 20-50 seconds which we obviouslydo not want.

Moreover we do not seem to need to be acquiring lock when removing layer as of the latest editition of autoconnect.

Also do not force autoconnect to connect after removal of layer. Instead mark it to be autonnected by the autoconnect thread on next tick.