ClipSettingsBar: Hide copy/paste, start/stop pos, reorder items, fix positioning
We don't need to show those copy/paste buttons here, and the start and
stop positions are done with direct interaction on the waveform based
wave editor.

The reordered items means we now have Loudness, Pitch, Speed Ratio, the
BPM settings, and length snapping in that order, so we can adjust this
more logically using the knobs k1, k2, k3, and bk in that order.

The positioning was previously done with a gridlayout, which becomes
awkward when things change and we need to redo column spans and the
like. So, metadata wants to be on a line of its own, and that's now the
case. Much nicer.

There is also now space here for adding in settings for filter control
for samplersynth... hmm...