Don't show progress dots for things with invalid progress IDs
This gets rid of that seemingly single stray dot at the start (it was
actually all the dots not currently in use, because... well, yeah, of
course it was).

Using the gain is a bit weird anyway, technically a thing can have no
gain and still be playing, so... yup, this makes better sense.