Add ClipGranulatorSettings component for handling CAS granular settings
This is not quite complete yet, as the position, spray, and scan options
are not implemented yet. The rest function, though, and allow for nicer
control of the CAS granular settings.

Note that the settings are not yet persisted (and we don't yet emit the
signal that would make that happen once implemented), so anything done
here would not persist through a restart.