Grainerator: Fix name, add pitch settings, split into three, add voice info
The settings panel now includes settings for individual grain pitch
shifting, as implemented in SamplerSynth's Granulator.

Further, we have a basic display of how many voices given interval and
spray settings are likely to produce, though we probably want to
visualise that a bit better (and also include a way of suggesting how
much polyphony those settings are able to support).

Also, the panel now has the correct name, and has been split into three
sections, called Notes, Voices, and Positions, which makes it easier to
navigate and also makes it less squeezed.