Autoconnect: Adapt to the new SamplerSynth lanes concept
This just brings us to the state we used to have. We will want to
connect and disconnect these lanes according to whether or not there is
a sample in a given slot, as well as assign those samples a lane
according to what they should be using (for 1-to-1 an individual lane,
for "standard" just all on lane 0). For now, though, things should work
as they used to.