Add ActionPickerPopup component for picking between a list of actions
This is intended to replace the popups with explicit buttons in them, so
we can adjust the look and feel of all of them simultaneously.

Likely need some more magic in there (a way to wrap things so if the
thing gets too tall, we still have space for the buttons), but for an
initial version, this'll do.