ActionPickerPopup: Fix problem with buttons staying lit after picking
Prior to this change, the buttons would (for whatever reason i just
don't get) stay highlighted once you'd picked one. For example,
alternating between picking a recording and removing the sample in the
sample picker popup, both of those buttons would remain highlighted,
until you pressed a finger down and moved outside the control. Just,
really not cricket. This uses the modifications in PlayGridButton to
make that control look more like a normal button, without the expensive
theming bits included. A touch annoying to have to use this to work
around what has to be a bug in qqc2.button, but... this works as
expected, so it'll do.