zynthian_engine : Try restarting engine if it fails to start
When cold booting, sometimes some engines fail to start up. Once an engine fails to start, zynthian does not try to restart which causes catastrophic cascading failures. When application starts up after a failed engine startup, it tries saving last state snapshot which excludes the stopped engine as it cannot find the engine instance and causes loss of data.

To circumvent this, try to restart engine in case of failure as a failsafe. To have a failsafe for the failsafe, restart the entire application if the engine fails to start even after 5 restarts.

Note : It was noticed that most of the times the FX engines fail to start on cold boot and not the synth ones. Also it never fails to start engines when just UI is restarted.