sketchpad: Add getter for channel recording filenames, honour do_countin
The first part is used to ensure we can get a useful filename for any
recordings we want to make for a given channel (such as when bouncing).

The second part is just to make it possible to not have a countin show
up when we start a recording, if we don't want one.