BouncePopup: Rework to using scheduling for recording and song mode for playing
This is basically working, but is showing me very clearly that we are
going to need to move the current playback start/stop logic into
c++, so we don't have to wait for the roundtripping of those
cross-thread signal/slot connections. In short, there is a gap at the
start of recordings, and that just is really not useable. We could work
around it, but it'd be more work than just doing it properly.