sketchpad.clip: Add metadataPatternJson property
This is supposed to be used when bouncing and unbouncing a sketch. If
bounced, we store the pattern directly (even if other midi data was
recorded), and use that for unbouncing. We retain the ability to apply
midi recorded when creating sketches in other ways (like when
recording), but this now means we end up with a fully correct unbounce,
instead of the occasionally slightly incorrect one that will happen when
recording the live midi events.