sketchpad: Adjust recorder logic to the new behind-the-scenes logic
More of the recording is controlled from c++, but because it's in the
timer, we will need to wait for those things to be done. Also, we should
be sort of gentle about it. Also we no longer need to stop the audio
recording here, so just leave that out.

Needs updated libzynthbox