stepsequencer: Don't highlight used notes in channel loop mode
Prior to this, we'd have the used notes in the selected part's pattern
highlighted, but what you could see at the bottom was the waveform, not
the pattern details, so... just looked a bit weird. Probably also need
to not show this for the used notes and whatnot, but this helps.