engine_fluidsynth : Populate jackname accordingly since there are now individual instance of fluidsynth engines
Count the previous number of fluidsynth engines available and populate jackname with it so it doesnt clash with existing engines.

Pass the engine name when creating fluidsynth otherwise jack automatically gives it a different name which cannot be retreived easily. Better to generate our own unique name and pass it instead