autoconnect: Rework routing setup to incorporate serial/one-to-one logic
This also changes the layout of the logic itself, which now more closely
follows our basic concepts (tracks, parts, and slots), and we now also
only connect things which actually have a sound source, resulting in a
much leaner looking graph (and less processing in simpler setups, which
always is nice). This is not quite complete yet, as we still need to
adjust some interaction points in the UI (most notably the channel send
knobs, but everything needs a quick look), and also we need to handle
the one-to-one situation, where we have individual send values for each
of the five channel sound slots.