sketchpad.channel: Add per-audiotype mixer settings to channel
This adds a full set of mixer settings to the channel data, and also
persists that information into the sketches. This ensures that when you
switch from e.g. synth to sketch mode, such as when you bounce a synth
track to a sketch, you don't then run a recording through the same
effects you just recorded.

Also adjust the UI to match the new logic