NewStuffPage: Dynamically create newStuffEngine, so reloading works
This also removes the pre-allocation of a fairly expensive object,
leaving a bit more space that we can then use for more generally useful
things (instead of just having a pile of newstuff engines sitting around
doing nothing). But primarily it means that when you back out and return
to the page, it will refresh the contents of the list, which we can't do
by just calling refresh on the engine, since that isn't exposed to qml
in kf5.