Sketchpad Picker: Tell people picking a community sketch that we'll copy it
The intent here is to ensure people are aware what is going to happen
when they pick a sketchpad from the community section. Since we
essentially treat them as a kind of template, we want to make sure that
is clear at load time (so people don't end up surprised).