stepsequencer: Slightly rework the pattern load/save popup options
The old load-sequence concept basically doesn't really work any longer.
It was previously based on the idea that a Sequence was those patterns
tied to a specific track. That changed several iterations of our
concepts ago, and it's just not a thing any longer. So, let's just
ignore that entirely, and just do single pattern load instead.