stepsequencer: Hide actions (and comment out the load/save dlg)
As we focus more tightly down on using sketches (with their inherent
previewability) as a sharing tool, we no longer really want to share
just the raw midi without it being embedded in a Sketch. So, hide the
load/save dialogue, and also don't actually offer the ability to
download the things.