sketchpad_channel : Heuristic to fix a problem with ZynAddSubFX
While switching channel_audio_type for a couple of times, ZynAddSubFX goes completely berserk and causes jack to become unresponsive and hence causing everything to go out-of-order. Testing suggests ZynAddSubFX does not like handling midi events it does not quite know about. During testing it was noticed that if ZynAddSubFX was disconnected before changing channel_audio_type and reconnected on complete it does not cause the aforementioned issue. Hence make sure to do the disconnect-connect dance when changing channel_audio_type only to the ZynAddSubFX synth if the track has one