NewStuffPage: Redo preview look'n'feel, add a set of useThis functionality
The useThis functionality allows the individual page to show a Use This
button on the page for installed items. The individual page can then
handle the signal fired by pressing that button, and act accordingly.
For example, you might have downloaded a bunch of sketches, and then
there's one you want to actually use now. Or you've downloaded a couple
of Sketchpads, and now you want to create something based on that. Or
you've downloaded some themes, and now you want to use one of those
themes (an example of maybe not wanting to go back upon using, as
described in the documentation).