segments model: Add insert_clip, ensure_split, fix missing signals
ensure_split is designed to ensure that there is a segment in the model
which ends at the given position. This is used by the insert_clip
function to allow the beat-wise precision positioning of a given clip
into a segments model.

Also fixed a couple of missing signals when creating new segments, which
is why the song duration would be out of whack when, well, creating new
segments, as opposed to changing the existing ones.