BouncePopup: Add full-sketch/all-track/single-part bounce logic
This covers a fair bit of work, and the concept is based around using
the song mode to play each of our loops in order, using the logic
originally created for the single-part bounce, but expanded to cover an
entire sketchpad if desired.

More testing will be wanted to ensure things work as they are supposed
to, but it feels pretty good so far.