TrackUnbouncer: Add new component for unbouncing full sketchpad Tracks
The concept is to allow for unbouncing an entire track's worth of
sketches, and covering the various potential corner cases that might
show up when doing so. Primarily, those corners will revolve around
unmatched sound setups when the bounce was originally created. We solve
this by not refusing to unbounce, but rather by offering the user the
options that the sketches include, and letting them pick for themselves.

The component is mostly complete, but requires further polish and is not
entirely ready for wide-spread testing just yet, but it also does show
the intent of the functionality. More work to be done, then, but also
there's no reason not to not push the work.