stepsequencer: Rework the listening logic somewhat
We no longer have an explicit listening state, but instead use what we
have in other places to perform that magic. It also results in much
faster and more stable feeling playing using the note pads, both during
and out of playback, which is lovely.

We also no longer have the press-to-show thing. It was frankly always
kind of awkward (as it would accidentally trigger if you were just
trying to play a thing for a while), and it is also just more
straightforward to show that same dialog with two taps (tap the note,
then tap the x in bar button)  than by tap-and-hold anyway.