BouncePopup: Entirely remove the start-recording delay
This used to be needed, but with various changes around the place,
things are controlled much sharply now, and it is no longer needed.
Leaving the variable there, with a comment describing the concept, but
we can probably get rid of that at some point if it turns out to just...
keep working.