zynthian.newstuff: Remove unneeded bits, add a connection checker
This was always something of a hack, and with how we have fundamentally
changed the way we handle the scrolling logic and such, this whole
multilayer thing just didn't make any sense any longer.

The connection checker is a basic function to ask to test whether
there's an internet connection (and specifically one that can get to the
store's location), and reporting back on that once the test is
completed. It's done in a thread like this to avoid high cost blocking
i/o calls.